gets to the point of returning ~48 bites, need to add authorization header
