+++ PubgStats/PubgStats/Base.lproj/Main.storyboard
@@ -21,7 +21,7 @@
                             <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="PUBG_Icon_Color" translatesAutoresizingMaskIntoConstraints="NO" id="P4N-CA-tQT">
                                 <rect key="frame" x="0.0" y="0.0" width="414" height="546"/>
                             </imageView>
-                            <textField opaque="NO" contentMode="scaleToFill" ambiguous="YES" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Enter Steam Username" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="Irm-ch-WjC">
+                            <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Enter Steam Username" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="Irm-ch-WjC">
                                 <rect key="frame" x="20" y="516" width="374" height="30"/>
                                 <nil key="textColor"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
@@ -52,6 +52,7 @@
                             <constraint firstAttribute="bottom" secondItem="P4N-CA-tQT" secondAttribute="bottom" constant="350" id="d2e-MC-Yvh"/>
                             <constraint firstAttribute="trailing" secondItem="P4N-CA-tQT" secondAttribute="trailing" id="dNQ-JK-7tP"/>
                             <constraint firstItem="d4x-Rd-bxr" firstAttribute="top" secondItem="P4N-CA-tQT" secondAttribute="bottom" constant="56" id="dWa-gD-tAI"/>
+                            <constraint firstItem="Irm-ch-WjC" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="472" id="koP-8M-Mwb"/>
                             <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="Irm-ch-WjC" secondAttribute="trailing" constant="20" id="sQo-XB-Enx"/>
                             <constraint firstItem="P4N-CA-tQT" firstAttribute="centerX" secondItem="8bC-Xf-vdC" secondAttribute="centerX" id="wln-Es-kcv"/>
                         </constraints>
